# src/main/resources/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:24110","description":"Generated server url"}],"paths":{"/api/v2/creators/{id}":{"get":{"tags":["creator-controller"],"operationId":"getCreatorById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelOut"}}}}}},"put":{"tags":["creator-controller"],"operationId":"updateCreator","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelOut"}}}}}},"delete":{"tags":["creator-controller"],"operationId":"deleteCreator","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/v1.0/stories":{"get":{"tags":["story-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Out"}}}}}}},"put":{"tags":["story-controller"],"operationId":"update","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"post":{"tags":["story-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}}},"/api/v1.0/notices":{"get":{"tags":["notice-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Out"}}}}}}},"put":{"tags":["notice-controller"],"operationId":"update_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"post":{"tags":["notice-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}}},"/api/v1.0/marks":{"get":{"tags":["mark-controller"],"operationId":"getAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Out"}}}}}}},"put":{"tags":["mark-controller"],"operationId":"update_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"post":{"tags":["mark-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}}},"/api/v2/creators":{"get":{"tags":["creator-controller"],"operationId":"getAllCreators","parameters":[{"name":"page","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","required":false,"schema":{"type":"string","default":"id::asc"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelOut"}}}}}},"post":{"tags":["creator-controller"],"operationId":"createCreator","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/In"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EntityModelOut"}}}}}}},"/api/v2/creators/firstnames/{firstname}":{"get":{"tags":["creator-controller"],"operationId":"getCreatorsByFirstname","parameters":[{"name":"firstname","in":"path","required":true,"schema":{"type":"string"}},{"name":"page","in":"query","required":false,"schema":{"minimum":0,"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelOut"}}}}}}},"/api/v1.0/stories/{id}":{"get":{"tags":["story-controller"],"operationId":"read","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"delete":{"tags":["story-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1.0/notices/{id}":{"get":{"tags":["notice-controller"],"operationId":"read_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"delete":{"tags":["notice-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}},"/api/v1.0/marks/{id}":{"get":{"tags":["mark-controller"],"operationId":"read_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Out"}}}}}},"delete":{"tags":["mark-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"In":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"maxLength":64,"minLength":2,"type":"string"},"password":{"maxLength":128,"minLength":8,"type":"string"},"firstname":{"maxLength":64,"minLength":2,"type":"string"},"lastname":{"maxLength":64,"minLength":2,"type":"string"}}},"EntityModelOut":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"Out":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"login":{"type":"string"},"password":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelOut":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"outList":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelOut"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}